requirements.txt
aio-pika==9.3.0
aiobotocore==2.7.0
aiohttp==3.8.6
aioitertools==0.11.0
aiormq==6.7.7
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.3
botocore==1.31.64
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.4
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fakeredis==2.20.0
fastapi==0.104.0
fastapi-mail==1.4.1
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
hiredis==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
pamqp==3.2.1
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
redis==5.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
wrapt==1.16.0
yarl==1.9.2
